впринципі все дороблено, лишилася телефонна версія і картинка у вчителя ну і міні помилки в frontend пошукати

+++ src/main/java/com/example/demo/dto/programe/LessonDTO.java
@@ -18,7 +18,7 @@ public class LessonDTO {
     private TeacherDTO teacher;
     private CourseDTO course;
     private LocalDateTime lessonTime;
-    private int duration;
+    private float duration;
     private String status;
 
 }
